Django Application/requirements.txt
altair==4.1.0
asgiref>=3.3.2,<4
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.1.1
bleach==3.3.0
blinker==1.4
cachetools==4.2.4
certifi==2024.6.2
chardet==4.0.0
click==8.0.4
cmake==3.18.4
colorama==0.4.4
cycler==0.10.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==4.3.1
django-rest-framework==3.14.0
dlib==19.22.0
docutils==0.16
entrypoints==0.4
enum-compat==0.0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
future==1.0.0
google==3.0.0
google-api-core==1.31.5
google-api-python-client==2.34.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.53.0
httplib2==0.22.0
idna==3.7
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.16.0
Jinja2==2.11.3
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab==3.0.16
jupyterlab-server==2.6.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.1
nbconvert==6.0.7
nbformat==5.1.3
notebook==6.4.10
numpy==1.19.5
opencv-python==4.5.5.64
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==4.21.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pydeck==0.6.2
Pygments==2.14.0
pyparsing==3.1.2
pyrsistent==0.18.0
python-dateutil==2.9.0
pytz==2022.6
pywin32==227
pywinpty==1.1.0
PyYAML==6.0.1
pyzmq==22.3.0
requests==2.27.1
rsa==4.9
s3transfer==0.5.2
Send2Trash==1.8.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
terminado==0.12.1
testpath==0.6.0
toml==0.10.2
toolz==0.12.0
torch==1.10.0
torchvision==0.11.1
tornado==6.1
traitlets==4.3.3
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.20
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==3.5.1